requirements.txt
torch~=1.8.1
torchtext>=0.10.0
pytorch-lightning>=1.3.5
code2seq==0.0.2
antlr4-python3-runtime~=4.9.2
setuptools>=52.0.0
tqdm~=4.32.2
numpy>=1.20.1
regex>=2019.11.1
omegaconf~=2.0.6
dataclasses~=0.6
requests~=2.25.1